added logic to check for default image before deleting instance from S3 bucket

+++ music/models.py
@@ -89,7 +89,8 @@ def delete_track_from_s3_bucket(sender, instance, **kwargs):
 
     client = boto3.client('s3', aws_access_key_id=aws_key_id, aws_secret_access_key=aws_secret_key)
     client.delete_object(Bucket=aws_bucket_name, Key=f'{instance.file}')
-    client.delete_object(Bucket=aws_bucket_name, Key=f'{instance.track_photo}')
+    if instance.track_photo!="track_photos/default.jpeg":
+        client.delete_object(Bucket=aws_bucket_name, Key=f'{instance.track_photo}')
 
 # delete video files and images for deleted track from AWS S3 bucket
 @receiver(post_delete, sender=Video)
@@ -100,4 +101,5 @@ def delete_video_from_s3_bucket(sender, instance, **kwargs):
 
     client = boto3.client('s3', aws_access_key_id=aws_key_id, aws_secret_access_key=aws_secret_key)
     client.delete_object(Bucket=aws_bucket_name, Key=f'{instance.file}')
-    client.delete_object(Bucket=aws_bucket_name, Key=f'{instance.video_photo}')
+    if instance.track_photo!="video_photos/default.jpeg":
+        client.delete_object(Bucket=aws_bucket_name, Key=f'{instance.video_photo}')
